--- index.js
@@ -1,4 +1,0 @@
-const { run } = require("@probot/adapter-github-actions");
-const app = require("./app");
-
-run(app);
