+++ index.ts
@@ -1,0 +1,4 @@
+const { run } = require("@probot/adapter-github-actions");
+const botApp = require("./app");
+
+run(botApp);
