+++ tsconfig.json
@@ -1,0 +1,25 @@
+/* To learn more about this file see: https://angular.io/config/tsconfig. */
+{
+    "compileOnSave": false,
+    "compilerOptions": {
+      "baseUrl": "../",
+      "outDir": "./dist",
+      "forceConsistentCasingInFileNames": true,
+      "resolveJsonModule": true,
+      "strict": false,
+      "noImplicitReturns": true,
+      "noFallthroughCasesInSwitch": true,
+      "sourceMap": true,
+      "declaration": false,
+      "downlevelIteration": true,
+      "experimentalDecorators": true,
+      "moduleResolution": "node",
+      "importHelpers": true,
+      "target": "es2017",
+      "module": "es2020",
+      "lib": [
+        "es2018",
+        "dom"
+      ]
+    }
+  }
